Fix wrong note in They All Laughed
Looks like this was an error in 'Jazz Fakebook' that got copied into Openbook.
I checked three other sources:

- A Frank Sinatra song book
- http://www.danmansmusic.com/members/notation/pdf/23040x.pdf
- http://xa.yimg.com/kq/groups/20774692/911263985/name/George+Gershwin+-+They+All+Laughed.pdf

They all have it as a b. Plus, it seem unlikely to have a c natural against the
c sharp in the chord (A7), which is what got me curious in the first place.